big update: implement PredicateBuilder to chain predicates
