+++ src/main/java/seedu/address/model/ModelManager.java
@@ -18,6 +18,7 @@ import seedu.address.model.tag.UniqueTagList;
 
 import java.time.LocalDateTime;
 import java.util.Set;
+import java.util.function.Predicate;
 import java.util.logging.Logger;
 
 /**
@@ -118,126 +119,14 @@ public class ModelManager extends ComponentManager implements Model {
     }
 
     @Override
-    public void updateFilteredEntryListByKeywords(Set<String> keywords){
-        updateFilteredPersonList(new PredicateExpression(new NameQualifier(keywords)));
+    public void updateFilteredEntryListPredicate(Predicate<Entry> predicate) {
+        updateFilteredPersonList(predicate);
     }
     
-    @Override
-    public void updateFilteredEntryListByStartDate(LocalDateTime startDate) {
-        updateFilteredPersonList(new PredicateExpression(new DateAfterQualifier(startDate)));
+    private void updateFilteredPersonList(Predicate<Entry> predicate) {
+        filteredPersons.setPredicate(predicate);
     }
+
     
-    @Override
-    public void updateFilteredEntryListByEndDate(LocalDateTime endDate) {
-        updateFilteredPersonList(new PredicateExpression(new DateBeforeQualifier(endDate)));
-    }
-
-    private void updateFilteredPersonList(Expression expression) {
-        filteredPersons.setPredicate(expression::satisfies);
-    }
-
-    //========== Inner classes/interfaces used for filtering ==================================================
-
-    interface Expression {
-        boolean satisfies(Entry person);
-        String toString();
-    }
-
-    private class PredicateExpression implements Expression {
-
-        private final Qualifier qualifier;
-
-        PredicateExpression(Qualifier qualifier) {
-            this.qualifier = qualifier;
-        }
-
-        @Override
-        public boolean satisfies(Entry person) {
-            return qualifier.run(person);
-        }
-
-        @Override
-        public String toString() {
-            return qualifier.toString();
-        }
-    }
-
-    interface Qualifier {
-        boolean run(Entry person);
-        String toString();
-    }
-
-    private class NameQualifier implements Qualifier {
-        private Set<String> nameKeyWords;
-
-        NameQualifier(Set<String> nameKeyWords) {
-            this.nameKeyWords = nameKeyWords;
-        }
-
-        @Override
-        public boolean run(Entry entry) {
-            return nameKeyWords.stream()
-                    .filter(keyword -> StringUtil.containsIgnoreCase(entry.getTitle().fullTitle, keyword))
-                    .findAny()
-                    .isPresent();
-        }
-
-        @Override
-        public String toString() {
-            return "name=" + String.join(", ", nameKeyWords);
-        }
-    }
-    
-    private class DateAfterQualifier implements Qualifier {
-        private LocalDateTime startDate;
-
-        DateAfterQualifier(LocalDateTime startDate) {
-            this.startDate = startDate;
-        }
-
-        // TODO: Change this when we introduce Events
-        @Override
-        public boolean run(Entry entry) {
-            // Don't include FloatingTasks, which have no deadline
-            if (entry.getClass().getSimpleName().equals("FloatingTask")) {
-                return false;
-            }
-             
-            // Deadline
-            Deadline deadline = (Deadline) entry;
-            return deadline.getDeadline().compareTo(startDate) >= 0;
-        }
-
-        @Override
-        public String toString() {
-            return "Due after: " + startDate.toString();
-        }
-    }
-    
-    private class DateBeforeQualifier implements Qualifier {
-        private LocalDateTime endDate;
-
-        DateBeforeQualifier(LocalDateTime endDate) {
-            this.endDate = endDate;
-        }
-
-        // TODO: Change this when we introduce Events
-        @Override
-        public boolean run(Entry entry) {
-            // Don't include FloatingTasks, which have no deadline
-            if (entry.getClass().getSimpleName().equals("FloatingTask")) {
-                return false;
-            }
-             
-            // Deadline
-            Deadline deadline = (Deadline) entry;
-            return deadline.getDeadline().compareTo(endDate) <= 0;
-        }
-
-        @Override
-        public String toString() {
-            return "Due before: " + endDate.toString();
-        }
-    }
 
 }
